refactor(nuxt): use compatibility version 4 with resets
