+++ src/config/modules/pwa.ts
@@ -1,7 +1,7 @@
 import type { DefineNuxtConfig } from 'nuxt/config'
 import colors from 'tailwindcss/colors'
 
-import { CACHE_VERSION, STYLE_DEFAULT_GRAY } from '../../utils/constants'
+import { CACHE_VERSION, STYLE_DEFAULT_GRAY } from '../../shared/utils/constants'
 
 const gray = colors[STYLE_DEFAULT_GRAY] // or gray, neutral, slate, stone
