+++ src/server/utils/auth.ts
@@ -3,7 +3,7 @@ import fs from 'node:fs'
 import type { H3Event } from 'h3'
 import { decodeJwt, jwtVerify, importSPKI } from 'jose'
 
-import { JWT_ALGORITHM } from '~/utils/constants'
+import { JWT_ALGORITHM } from '~/shared/utils/constants'
 
 // const configPostgraphileJwtPublicKeyPath =
 //   process.env.POSTGRAPHILE_JWT_PUBLIC_KEY_FILE || ''
