+++ src/server/utils/ical.ts
@@ -2,7 +2,7 @@ import DOMPurify from 'isomorphic-dompurify'
 import ical, { ICalCalendarMethod, ICalEventStatus } from 'ical-generator'
 import mustache from 'mustache'
 
-import { getTextFromHtml } from '~/utils/text'
+import { getTextFromHtml } from '~/shared/utils/text'
 import type {
   ContactItemFragment,
   EventItemFragment,
