+++ src/server/utils/testing.ts
@@ -1,5 +1,5 @@
 import type { H3Event } from 'h3'
-import { TESTING_COOKIE_NAME } from '~/utils/constants'
+import { TESTING_COOKIE_NAME } from '~/shared/utils/constants'
 
 export const isTestingServer = (event?: H3Event) => {
   const isTestingByRuntimeConfig = useRuntimeConfig().public.vio.isTesting
