+++ src/server/utils/timezone.ts
@@ -1,6 +1,6 @@
 import type { H3Event } from 'h3'
 
-import { TIMEZONE_COOKIE_NAME } from '~/utils/constants'
+import { TIMEZONE_COOKIE_NAME } from '~/shared/utils/constants'
 
 // TODO: rename to `getTimezone` (https://github.com/nuxt/cli/issues/266)
 export const getTimezoneServer = async (event: H3Event) => {
