+++ src/shared/utils/development.ts
@@ -1,0 +1,6 @@
+import seedrandom from 'seedrandom'
+
+export const getHeroImageName = (seed: string) => {
+  const images = ['conference', 'crowd', 'party', 'workshop']
+  return images[Math.floor(seedrandom(seed)() * images.length)]
+}
