+++ src/shared/utils/model.ts
