+++ src/shared/utils/networking.ts
