+++ src/shared/utils/performance.ts
@@ -1,0 +1,2 @@
+export const zalgo = async <T>(maybePromise: T) =>
+  import.meta.server ? await maybePromise : maybePromise
