+++ src/utils/auth.ts
@@ -3,7 +3,7 @@ import { consola } from 'consola'
 import { type H3Event, setCookie } from 'h3'
 import { decodeJwt } from 'jose'
 
-import { JWT_NAME } from './constants'
+import { JWT_NAME } from '../shared/utils/constants'
 import { authenticateMutation } from '~/gql/documents/mutations/account/accountAuthenticate'
 import { jwtRefreshMutation } from '~/gql/documents/mutations/account/accountJwtRefresh'
